requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
pydantic==2.8.2
email_validator==2.2.0
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4 